Organize back to full build
